@@ -6,7 +6,7 @@ description: Demonstrates how to use the sound_stream plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
@@ -22,8 +22,8 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-  web_socket_channel: ^1.1.0
+  cupertino_icons: ^1.0.6
+  web_socket_channel: ^2.4.4
 
 dev_dependencies:
   flutter_test:
