@@ -20,10 +20,11 @@ class _MyAppState extends State<MyApp> {
   List<Uint8List> _micChunks = [];
   bool _isRecording = false;
   bool _isPlaying = false;
+  bool _useSpeaker = false;
 
-  StreamSubscription _recorderStatus;
-  StreamSubscription _playerStatus;
-  StreamSubscription _audioStream;
+  StreamSubscription? _recorderStatus;
+  StreamSubscription? _playerStatus;
+  StreamSubscription? _audioStream;
 
   @override
   void initState() {
@@ -87,18 +88,33 @@ class _MyAppState extends State<MyApp> {
         appBar: AppBar(
           title: const Text('Plugin example app'),
         ),
-        body: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceAround,
+        body: Column(
+          crossAxisAlignment: CrossAxisAlignment.center,
           children: [
-            IconButton(
-              iconSize: 96.0,
-              icon: Icon(_isRecording ? Icons.mic_off : Icons.mic),
-              onPressed: _isRecording ? _recorder.stop : _recorder.start,
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceAround,
+              children: [
+                IconButton(
+                  iconSize: 96.0,
+                  icon: Icon(_isRecording ? Icons.mic_off : Icons.mic),
+                  onPressed: _isRecording ? _recorder.stop : _recorder.start,
+                ),
+                IconButton(
+                  iconSize: 96.0,
+                  icon: Icon(_isPlaying ? Icons.pause : Icons.play_arrow),
+                  onPressed: _isPlaying ? _player.stop : _play,
+                ),
+              ],
             ),
             IconButton(
               iconSize: 96.0,
-              icon: Icon(_isPlaying ? Icons.pause : Icons.play_arrow),
-              onPressed: _isPlaying ? _player.stop : _play,
+              icon: Icon(_useSpeaker ? Icons.headset_off : Icons.headset),
+              onPressed: () {
+                setState(() {
+                  _useSpeaker = !_useSpeaker;
+                  _player.usePhoneSpeaker(_useSpeaker);
+                });
+              },
             ),
           ],
         ),
