@@ -49,11 +49,11 @@ void writeWaveFileHeader(File out, int totalAudioLen, int longSampleRate,
 }
 
 main() {
-  Timer timer;
-  File file, tmp;
-  int dataSize;
+  Timer? timer;
+  File? file, tmp;
+  int dataSize = 0;
   WebSocket.connect('ws://localhost:$_PORT').then((WebSocket ws) {
-    if (ws?.readyState == WebSocket.open) {
+    if (ws.readyState == WebSocket.open) {
       ws.listen(
         (data) {
           if (timer == null) {
@@ -61,16 +61,16 @@ main() {
             final fileName = '$timestamp.wav';
             file = File(fileName);
             tmp = File('$fileName.tmp');
-            file.openWrite();
-            tmp.openWrite();
+            file!.openWrite();
+            tmp!.openWrite();
             dataSize = 0;
             timer = Timer(const Duration(seconds: 5), () {
               print('Timeout!');
-              writeWaveFileHeader(file, dataSize, _sampleRate, 1,
+              writeWaveFileHeader(file!, dataSize, _sampleRate, 1,
                   _RECORDER_BPP * _sampleRate ~/ 8);
-              file.writeAsBytesSync(tmp.readAsBytesSync(),
+              file!.writeAsBytesSync(tmp!.readAsBytesSync(),
                   flush: true, mode: FileMode.append);
-              tmp.deleteSync();
+              tmp!.deleteSync();
               tmp = null;
               file = null;
               timer = null;
